+++ R/bigquery-dataset.R
@@ -18,12 +18,16 @@ bqDatasetExists <- function(dataset = bqDefaultDataset(),
 bqCreateDataset <- function(dataset = bqDefaultDataset(),
                             project = bqDefaultProject()) {
   bqAuth()
-  bq_dataset_create(
-    bq_dataset(
-      project = project,
-      dataset = dataset
+  if (!bqDatasetExists(dataset, project)) {
+    bq_dataset_create(
+      bq_dataset(
+        project = project,
+        dataset = dataset
+      )
     )
-  )
+  } else {
+    message("Dataset does already exist: ", dataset)
+  }
 }
 
 #' @rdname bqDataset
@@ -36,15 +40,20 @@ bqDeleteDataset <- function(dataset = bqDefaultDataset(),
                             delete.contents = FALSE) {
   bqAuth()
 
-  assert_that(!bqProtectedDataset(dataset, project))
+  if (bqDatasetExists(dataset, project)) {
+    assert_that(!bqProtectedDataset(dataset, project))
 
-  bq_dataset_delete(
-    bq_dataset(
-      project = project,
-      dataset = dataset
-    ),
-    delete_contents = delete.contents
-  )
+    bq_dataset_delete(
+      bq_dataset(
+        project = project,
+        dataset = dataset
+      ),
+      delete_contents = delete.contents
+    )
+  } else {
+    message("Dataset does not exist and will not be deleted: ", dataset)
+  }
+
 }
 
 #' Gets list of tables for a given dataset
@@ -82,4 +91,20 @@ bqProtectedDataset <- function(dataset, project) {
   else {
     FALSE
   }
+}
+
+#' Copies all tables from one dataset to another
+#' through schema definitions
+#'
+#' @export
+#' @param from dataset object created by bigrquery::bq_dataset()
+#' @param to dataset object created by bigrquery::bq_dataset()
+bqCopyDatasetSchema <- function(from, to) {
+
+  tables <- bq_dataset_tables(from)
+  lapply(tables, function(x) {
+    to <- bq_table(to$project, to$dataset, x$table)
+    bqCopyTableSchema(x, to)
+  })
+
 }