+++ R/bigquery-partition.R
@@ -104,10 +104,18 @@ bqPartitionDatesSql <- function(table) {
   }
 }
 
+#' Functions to update existing partitions in the target table
+#'
 #' @description `bqRefreshPartitionData` updates existing partitions in the target table
 #'
-#' @rdname bqPartition
+#' @rdname bqRefreshPartitionData
 #' @export
+#' @param table destination partition table where results of the query will be saved
+#' @param file path to the sql file that will be used for the transformation
+#' @param ...  parameters that will be passed via `sprintf` to build dynamic SQL.
+#'    partition date will be always passed first in format `yyyymmdd`
+#'    followed by arguments in `...`
+#' @inheritParams bqCreateTable
 bqRefreshPartitionData <- function(table,
                                    file,
                                    ...,
@@ -134,29 +142,36 @@ bqRefreshPartitionData <- function(table,
 #' Functions to transforms partitioned data form one table to another
 #'
 #' @description `bqTransformPartition` creates new partitions for the missing dates
-#' @rdname bqPartition
+#' @rdname bqTransformPartition
 #' @export
 #' @param table destination partition table where results of the query will be saved
 #' @param file path to the sql file that will be used for the transformation
 #' @param ...  parameters that will be passed via `sprintf` to build dynamic SQL.
 #'    partition date will be always passed first in format `yyyymmdd`
 #'    followed by arguments in `...`
+#' @param missing.dates dates for which to run this function for
+#' @param priority Default to INTERACTIVE
+#' @param use.legacy.sql Defaults to env variable if specified
 #' @inheritParams bqCreateTable
 bqTransformPartition <- function(table,
                                  file,
                                  ...,
+                                 missing.dates = NULL,
                                  priority = "INTERACTIVE",
                                  use.legacy.sql = bqUseLegacySql()) {
+
   existing.dates <- bqExistingPartitionDates(table)
   start.date <- bqStartDate(unset = "2017-01-01")
   end.date <- bqEndDate()
 
-  missing.dates <- getMissingDates(
-    start.date,
-    end.date,
-    existing.dates,
-    "%Y-%m-%d"
-  )
+  if (missing(missing.dates) || is.null(missing.dates)) {
+    missing.dates <- getMissingDates(
+      bqStartDate(),
+      bqEndDate(),
+      existing.dates,
+      "%Y-%m-%d"
+    )
+  }
 
   jobs <- lapply(missing.dates, function(d) {
     partition <- gsub("-", "", d)