+++ tests/testthat/test.bq.table.R
@@ -20,6 +20,8 @@ test_that("table is created from schema file", {
   expected.partition <- TRUE
 
   expect_s3_class(res, "bq_table")
+  meta <- bq_table_meta(res)
+  expect_equal(meta$timePartitioning$type, "DAY")
 })
 
 test_that("correct schema is generated for a given dataset", {
@@ -56,4 +58,17 @@ test_that("Table can be patched", {
     label = "Patch fails if field removal is attempted."
   )
 
+})
+
+test_that("Table with cluster can be created", {
+  res <- bqInitiateTable(
+    table = "test_table_cluster",
+    schema.file = "bq-table-schema.json",
+    partition = TRUE,
+    clustering = list(fields = list("first_name"))
+  )
+
+  expect_s3_class(res, "bq_table")
+  meta <- bq_table_meta(res)
+  expect_equal(meta$clustering$fields[[1]], "first_name")
 })